+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Now `HashFileReader` and `HashFileWriter` support `pathlib.Path`.
+
 ### Configuration
 
 - Migrated build system from `setuptools` to `hatch`.